+++ src/api/router.ts
@@ -6,6 +6,7 @@ import { performanceRouter } from './app/performance';
 import { reportsRouter } from './app/reports';
 import { usersRouter } from './app/users';
 import { workerRouter } from './app/worker';
+import { clientRestProcedures } from './public/client-rest';
 import { publicApiRouter } from './public/router';
 import { router } from './trpc';
 
@@ -19,6 +20,7 @@ export const appRouter = router({
     performance: performanceRouter,
     users: usersRouter,
     worker: workerRouter,
+    ...clientRestProcedures,
 });
 
 export type AppRouter = typeof appRouter;